quantum for folsom
This commit contains changes that I made to get quantum working with
folsom. While these changes do not lead to the fully automated install
of a fully functional quantum environment, they are much closer, with only
a few minor issues with floating ip routing and the metadata server left.

This commit also updates the nova tests to be able to use quantum. Again though,
it does not work 100%.